modified:   build.sh

+++ build.sh
@@ -6,18 +6,22 @@ then
 	echo "--- Executing build process ---"
 else
 	echo "input: $1"
-	echo "Usage: ./build.sh <path to reqT.jar>"
+	echo "Usage: ./build.sh <path to reqT.jar> <options>"
+	echo "Options:"
+	echo " -r	Rebuild tex from scala" 
 	exit
 fi
-
-java -jar $1 -i build.scala
-for f in docs/krav/*.tex;
-do
-	echo "Removing eventual chapter-shite and changing section to subsubsection in $f"
-	sed '/chapter/d' $f | cat > $f.tmp
-	sed 's/section/subsubsection/g' $f.tmp | cat > $f
-	rm $f.tmp
-done
+if [[$2 == "-r"]]
+then
+	java -jar $1 -i build.scala
+	for f in docs/krav/*.tex;
+	do
+		echo "Removing eventual chapter-shite and changing section to subsubsection in $f"
+		sed '/chapter/d' $f | cat > $f.tmp
+		sed 's/section/subsubsection/g' $f.tmp | cat > $f
+		rm $f.tmp
+	done
+fi
 
 cd docs/
 pdflatex systemRequirements.tex
